+++ src/main/java/org/jointheleague/features/student/first_feature/CurrentWeather.java
@@ -1,0 +1,41 @@
+package org.jointheleague.features.student.first_feature;
+
+import javax.annotation.Generated;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+@Generated("jsonschema2pojo")
+public class CurrentWeather {
+
+	@SerializedName("location")
+	@Expose
+	private Location location;
+	@SerializedName("current")
+	@Expose
+	private Current current;
+
+	public Location getLocation() {
+		return location;
+	}
+
+	public void setLocation(Location location) {
+		this.location = location;
+	}
+
+	public Current getCurrent() {
+		return current;
+	}
+
+	public void setCurrent(Current current) {
+		this.current = current;
+	}
+
+	@Override
+	public String toString() {
+		return "CurrentWeather [location=" + location + ", current=" + current + "]";
+	}
+
+	
+	
+}
